requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
numpy==1.26.2
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2
pytest-cov==4.1.0
opencv-python-headless==4.8.1.78
